@@ -3,14 +3,19 @@
 namespace App\Controller\Admin;
 
 use App\Entity\User;
-use Doctrine\Common\Collections\ArrayCollection;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\ArrayField;
-use EasyCorp\Bundle\EasyAdminBundle\Field\AssociationField;
+use EasyCorp\Bundle\EasyAdminBundle\Field\IdField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
+use Doctrine\ORM\EntityManagerInterface;
 
 class UserCrudController extends AbstractCrudController
 {
+    public function __construct(
+        private UserPasswordHasherInterface $passwordHasher
+    ) {}
+
     public static function getEntityFqcn(): string
     {
         return User::class;
@@ -19,10 +24,37 @@ class UserCrudController extends AbstractCrudController
     public function configureFields(string $pageName): iterable
     {
         return [
-            TextField::new('userIdentifier'),
+            IdField::new('id')->hideOnForm(),
+            TextField::new('username'),
             ArrayField::new('roles'),
-            TextField::new('plainPassword')->onlyOnForms(),
+            TextField::new('plainPassword')
+                ->setLabel('Mot de passe')
+                ->setRequired(false)
+                ->onlyOnForms()
+                ->setHelp('Laissez vide pour ne pas changer le mot de passe')
+                ->setFormTypeOption('attr', ['type' => 'password']),
         ];
     }
 
+    public function persistEntity(EntityManagerInterface $entityManager, $entityInstance): void
+    {
+        $this->hashPassword($entityInstance);
+        parent::persistEntity($entityManager, $entityInstance);
+    }
+
+    public function updateEntity(EntityManagerInterface $entityManager, $entityInstance): void
+    {
+        $this->hashPassword($entityInstance);
+        parent::updateEntity($entityManager, $entityInstance);
+    }
+
+    private function hashPassword(User $user): void
+    {
+        $plainPassword = $user->getPlainPassword();
+        
+        if ($plainPassword !== null && $plainPassword !== '') {
+            $hashedPassword = $this->passwordHasher->hashPassword($user, $plainPassword);
+            $user->setPassword($hashedPassword);
+        }
+    }
 }
