@@ -4,8 +4,8 @@ namespace App\Controller;
 
 use App\Repository\BankRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Security\Core\Security;
-
+use Symfony\Bundle\SecurityBundle\Security;
+use App\Entity\User;
 class BanksByUserController extends AbstractController
 {
     public function __construct(protected BankRepository $repository, private readonly Security $security) {
@@ -13,8 +13,10 @@ class BanksByUserController extends AbstractController
 
     public function __invoke(): ?array
     {
-        return $this->repository->findByUser($this->security->getUser());
+        $user = $this->security->getUser();
+        if (!$user instanceof User) {
+            return null;
+        }
+        return $this->repository->findByUser($user);
     }
-
-
 }
