@@ -6,7 +6,7 @@ use App\Entity\Bank;
 use App\Entity\User;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
-use Symfony\Component\Security\Core\User\UserInterface;
+
 
 /**
  * @method Bank|null find($id, $lockMode = null, $lockVersion = null)
@@ -54,7 +54,7 @@ class BankRepository extends ServiceEntityRepository
      * @param User $user
      * @return Bank[]|null
      */
-    public function findByUser(UserInterface $user): ?array
+    public function findByUser(User $user): ?array
     {
         return $this->createQueryBuilder('b')
             ->join('b.charges', 'c')
