@@ -4,13 +4,12 @@ namespace App\State;
 
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\State\ProviderInterface;
-use Symfony\Component\Security\Core\Security;
+use Symfony\Bundle\SecurityBundle\Security;
 
 class UserVerifyProvider implements ProviderInterface
 {
     public function __construct(private readonly Security $security)
     {
-        
     }
 
     public function provide(Operation $operation, array $uriVariables = [], array $context = []): object|array|null
